Move TypeLoadException to shared (#23238)
* fix coding-style

* shorter form for TypeName property

* move coreclr-specific ctor to TypeLoadException
